core, mcp, openai: clear pedantic lints and tighten hot paths
core: ToolImpl::definition now borrows the cached definition instead of
returning a clone, so registering a tool no longer copies its whole
argument schema; Tools::insert reads the name and description off the
boxed tool rather than taking a second copy of the definition.

mcp: flatten client/client.rs and server/server.rs into client.rs and
server.rs (module inception); export McpClient, which the crate docs
referenced but which was unreachable; deduplicate McpServer::stdio
against McpServer::new and make new public, since a custom transport is
a real use case; give write_message's message a Sync bound and take the
tool table rather than the whole server in handle_call_tool, so the
server futures are Send; hand run_service the receiver by reference and
the connection by value.

openai: split the Responses stream body into a ResponsesStreamState with
an apply() step function, so event handling is an ordinary readable
function rather than 150 lines inside a stream! macro; extract the chat
tool-call accumulation, drain, and SSE error probe; extract the Responses
request builder; group the built-in tool flags into BuiltinTools; rename
ResponsesInputContent variants off their shared Input prefix with
explicit serde renames; type OpenAIFile::created_at as u64.

+++ openai/src/files.rs
@@ -70,7 +70,7 @@ pub struct OpenAIFile {
     /// Size of the file in bytes.
     pub bytes: u64,
     /// Unix timestamp when the file was created.
-    pub created_at: i64,
+    pub created_at: u64,
     /// Name of the file.
     pub filename: String,
     /// Purpose of the file.
@@ -87,7 +87,7 @@ impl OpenAIFile {
     /// Get the creation time as `SystemTime`.
     #[must_use]
     pub fn created(&self) -> SystemTime {
-        UNIX_EPOCH + Duration::from_secs(self.created_at as u64)
+        UNIX_EPOCH + Duration::from_secs(self.created_at)
     }
 
     /// Check if the file is ready for use.
